Return the content itself, not an element
The third element in data contains the html when doing a call
to `#request` which  was returned by `.validate_binary_output`.
This works until `#response`is called where the meaningful data
is the whole array instead.

Finding the right element to return is `Atos` class's burden not
the binary validator anyway.

I suspect this bug was introduced in a refactoring.